# ENH: remove memory addresses from param repr()