The `pandas` benchmark suite sometimes uses functions as `params`, which largely works fine. The one issue is that the `repr()` of a function includes its memory address and can change over subsequent `asv` invocations (for normal functions) or even within a single `asv` run (for `lambda` functions). This change in `repr()` impacts the ability to track and compare a benchmark over time, leading to broken `asv compare` functionality like below:

```
Benchmarks that have stayed the same:

       before           after         ratio
     [24ab22f7]       [d7cef344]
              n/a          997±3μs      n/a  ctors.SeriesConstructors.time_series_constructor(<function SeriesConstructors.<lambda> at 0x7f0deb0800d0>, False)
              n/a         1.05±0ms      n/a  ctors.SeriesConstructors.time_series_constructor(<function SeriesConstructors.<lambda> at 0x7f0deb0800d0>, True)
      1.04±0.01ms              n/a      n/a  ctors.SeriesConstructors.time_series_constructor(<function SeriesConstructors.<lambda> at 0x7fd9d0bbc598>, False)
      1.06±0.02ms              n/a      n/a  ctors.SeriesConstructors.time_series_constructor(<function SeriesConstructors.<lambda> at 0x7fd9d0bbc598>, True)
```

This PR modifies `repr()` iff it appears to include a memory address, allowing proper comparison over time. The new output looks like this after a re-run:

```
       before           after         ratio
     [24ab22f7]       [d7cef344]
          995±7μs      1.03±0.02ms     1.04  ctors.SeriesConstructors.time_series_constructor(<function <lambda>>, False)
      1.05±0.01ms      1.09±0.05ms     1.03  ctors.SeriesConstructors.time_series_constructor(<function <lambda>>, True)

```

One unaddressed issue is that any suite with multiple `lambda` functions will now have name collisions in `params`. I intend to address this in `pandas` by creating normal named functions in the relevant suites. Alternative solutions would be to append distinct numbers or hashes to the `repr()`s, or to warn or raise if a user passes multiple `lambda` functions.