@qwhelan 2018-11-28
I went with raising an exception on non-unique `repr()` approach, as it doesn't seem like having `lambda_1` through `lambda_n` will ever be that useful in the UI.

Here's the output on `pandas` without my patch (`numpy` and `scipy` don't need any changes):
```
$ asv dev
· Discovering benchmarks
·· Error running /home/chris/anaconda3/bin/python /home/chris/projects/asv/asv/benchmark.py discover /home/chris/projects/pandas/asv_bench/benchmarks /tmp/tmpom4ey47w/result.json
   STDOUT -------->

   STDERR -------->
   Traceback (most recent call last):
     File "/home/chris/projects/asv/asv/benchmark.py", line 1230, in <module>
       main()
     File "/home/chris/projects/asv/asv/benchmark.py", line 1223, in main
       commands[mode](args)
     File "/home/chris/projects/asv/asv/benchmark.py", line 932, in main_discover
       list_benchmarks(benchmark_dir, fp)
     File "/home/chris/projects/asv/asv/benchmark.py", line 917, in list_benchmarks
       for benchmark in disc_benchmarks(root):
     File "/home/chris/projects/asv/asv/benchmark.py", line 825, in disc_benchmarks
       class_attr)
     File "/home/chris/projects/asv/asv/benchmark.py", line 771, in _get_benchmark
       return cls(name, func, sources)
     File "/home/chris/projects/asv/asv/benchmark.py", line 495, in __init__
       Benchmark.__init__(self, name, func, attr_sources)
     File "/home/chris/projects/asv/asv/benchmark.py", line 420, in __init__
       dupe_str))
   ValueError: Benchmark ctors.SeriesConstructors.time_series_constructor has duplicate string representations for param data_fmt:
   <function SeriesConstructors.<lambda>>: 7 copies

·· Failed to build the project and import the benchmark suite.
```