@qwhelan 2018-12-02
@pv Sorry, forgot to update that comment - I ended up switching approaches due to finding existing cases where it didn't make sense to force a `repr()` change. Here's current pandas output:
```
$ asv dev -b ctors.SeriesConstructors
· Discovering benchmarks
· Running 1 total benchmarks (1 commits * 1 environments * 1 benchmarks)
[  0.00%] ·· Benchmarking existing-py_home_chris_anaconda3_bin_python
[ 50.00%] ··· ctors.SeriesConstructors.time_series_constructor                                                                                                          ok
[ 50.00%] ··· ========================================================== ========== ==========
              --                                                               with_index
              ---------------------------------------------------------- ---------------------
                                       data_fmt                            False       True
              ========================================================== ========== ==========
               <function SeriesConstructors.<lambda> at 0x7fa7f0beba60>   1.49±0ms   1.81±0ms
                                         list                             3.12±0ms   3.33±0ms
               <function SeriesConstructors.<lambda> at 0x7fa7f0bebae8>   2.72±0ms   2.99±0ms
               <function SeriesConstructors.<lambda> at 0x7fa7f0bebb70>   6.91±0ms   7.37±0ms
               <function SeriesConstructors.<lambda> at 0x7fa7f0bebbf8>   2.95±0ms   3.03±0ms
               <function SeriesConstructors.<lambda> at 0x7fa7f0bebc80>   2.86±0ms   3.11±0ms
               <function SeriesConstructors.<lambda> at 0x7fa7f0bebd08>   2.74±0ms   3.08±0ms
               <function SeriesConstructors.<lambda> at 0x7fa7f0bebd90>   2.86±0ms   3.07±0ms
              ========================================================== ========== ==========
```
And here's the new output:
```
$ asv dev -b ctors.SeriesConstructors
· Discovering benchmarks
· Running 1 total benchmarks (1 commits * 1 environments * 1 benchmarks)
[  0.00%] ·· Benchmarking existing-py_home_chris_anaconda3_bin_python
[ 50.00%] ··· ctors.SeriesConstructors.time_series_constructor                                                                                                          ok
[ 50.00%] ··· ============================================ ========== ==========
              --                                                 with_index
              -------------------------------------------- ---------------------
                                data_fmt                     False       True
              ============================================ ========== ==========
               <function SeriesConstructors.<lambda>> (0)   1.41±0ms   1.81±0ms
                                  list                      3.16±0ms   3.40±0ms
               <function SeriesConstructors.<lambda>> (1)   2.77±0ms   2.96±0ms
               <function SeriesConstructors.<lambda>> (2)   6.90±0ms   7.48±0ms
               <function SeriesConstructors.<lambda>> (3)   2.88±0ms   2.94±0ms
               <function SeriesConstructors.<lambda>> (4)   2.94±0ms   3.13±0ms
               <function SeriesConstructors.<lambda>> (5)   2.88±0ms   3.04±0ms
               <function SeriesConstructors.<lambda>> (6)   4.02±0ms   2.98±0ms
              ============================================ ========== ==========
```